+++ start_mcp_inspector.sh
@@ -3,6 +3,8 @@ echo "Killing any existing FastMCP inspector processes..."
 lsof -ti :6277 | xargs kill -9
 lsof -ti :6274 | xargs   kill -9
 
+export DANGEROUSLY_OMIT_AUTH=true
+
 echo "Starting eBayAPI-MCP-Server inspector..."
 npx @modelcontextprotocol/inspector --config /Users/petetreadaway/Projects/ebay-mcp-server/MCP_Inspector_Config.json --server ebayAPI-mcp-server